+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
 ## 1.0.2
+- removed 17 chars hostname length limit
+- added HW1 and HW5 support
+- utility that creates TS format compatible firmware
+- logging to TS API calls
 - added flask_migrate
+- added PID parameters
+- added API to be called from HA integration
 
 ## 1.0.1